Move binaries into cmd folder
