+++ build.sh
@@ -10,10 +10,10 @@ go get -u github.com/golang/dep/cmd/dep
 dep ensure
 
 # Build the service
-cd controller && go build -a -o kubernetes-vault
+go build -a -o cmd/controller/kubernetes-vault ./cmd/controller/
 
 # Build the init container
-cd ../init && go build -a -o kubernetes-vault-init
+go build -a -o cmd/init/kubernetes-vault-init ./cmd/init/
 
 # Build the sample-app container
-cd ../sample-app && go build -a -o sample-app
+go build -a -o cmd/sample-app/sample-app ./cmd/sample-app/
