+++ cmd/controller/Dockerfile
@@ -7,15 +7,14 @@ COPY . .
 RUN apk --update add gcc git musl-dev \
  && go get -u github.com/Masterminds/glide \
  && glide install --strip-vendor \
- && cd controller \
- && go build -o kubernetes-vault
+ && go build -o kubernetes-vault ./cmd/controller
 
 FROM alpine:latest
 LABEL maintainer Francis Chuang <francis.chuang@boostport.com>
 
 RUN mkdir -p /kubernetes-vault
 
-COPY --from=builder /go/src/github.com/Boostport/kubernetes-vault/controller/kubernetes-vault /bin/kubernetes-vault
-COPY --from=builder /go/src/github.com/Boostport/kubernetes-vault/controller/docker-entrypoint.sh /usr/local/bin/docker-entrypoint.sh
+COPY --from=builder /go/src/github.com/Boostport/kubernetes-vault/kubernetes-vault /bin/kubernetes-vault
+COPY --from=builder /go/src/github.com/Boostport/kubernetes-vault/cmd/controller/docker-entrypoint.sh /usr/local/bin/docker-entrypoint.sh
 
 ENTRYPOINT ["docker-entrypoint.sh"]
