+++ cmd/controller/docker-entrypoint.sh
