+++ cmd/init/Dockerfile
@@ -7,12 +7,11 @@ COPY . .
 RUN apk --update add gcc git musl-dev \
  && go get -u github.com/Masterminds/glide \
  && glide install --strip-vendor \
- && cd sample-app \
- && go build -o sample-app
+ && go build -o kubernetes-vault-init ./cmd/init
 
 FROM alpine:latest
 LABEL maintainer Francis Chuang <francis.chuang@boostport.com>
 
-COPY --from=builder /go/src/github.com/Boostport/kubernetes-vault/sample-app/sample-app /sample-app
+COPY --from=builder /go/src/github.com/Boostport/kubernetes-vault/kubernetes-vault-init /kubernetes-vault-init
 
-ENTRYPOINT ["/sample-app"]
+ENTRYPOINT ["/kubernetes-vault-init"]
